pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cli-session-log"
version = "0.1.0"
description = "Manage CLI sessions with conversation logs and task tracking"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["cli", "session", "log", "markdown", "obsidian"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Utilities",
]
dependencies = [
    "pyyaml>=6.0",
    "filelock>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
session-log = "cli_session_log.cli:main"

[project.urls]
Homepage = "https://github.com/thedomainai/cli-session-log"
Repository = "https://github.com/thedomainai/cli-session-log"
Issues = "https://github.com/thedomainai/cli-session-log/issues"

[tool.hatch.build.targets.wheel]
packages = ["cli_session_log"]
